content: 2e: chapter3: Fix typos at The Call Stack
After the execution returns to the line in function a() that called function b(). Next it calls the function d()  which prints 'd() starts' and 'd() returns' and this function doesn't call any other function. So, now it returns to the line in function a() that called function (d), then it prints 'a() returns' and then the main function call a() from the end of the program ends.

The chapter says:

1. The execution continues to the next line in the b() function, which is a call to d().
2. It just displays d() starts and d() returns before returning to the line in b() that called it.
3. Since b() contains no other code, the execution returns to the line in a() that called b().